Move components to different files
